# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MCG Building Violation Management System API
  description: REST API consumed by the MCG web portal module and the MCG HARYANA mobile app (building-violations module).
  version: 2.0.0
paths:
...
  /building-violations/api/inspections/tasks/{pk}/:
    get:
      tags:
      - building-violations
      summary: Task Detail
      operationId: task_detail_building_violations_api_inspections_tasks__pk___get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: integer
          title: Pk
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - building-violations
      summary: Task Update
      operationId: task_update_building_violations_api_inspections_tasks__pk___patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: integer
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TaskUpdateIn:
      properties:
        instructions:
          anyOf:
          - type: string
          - type: 'null'
          title: Instructions
        priority:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        due_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Due At
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        owner_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Name
        owner_mobile:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Mobile
      type: object
      title: TaskUpdateIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /building-violations/api/auth/otp/verify/
servers:
- url: /
  description: Same origin (the module is mounted at /building-violations/)